+++ src/solc_0.8/StarterPack/StarterPackV2.sol
@@ -35,17 +35,17 @@ contract StarterPackV2 is PurchaseValidator, ERC2771Handler {
     event Purchase(address indexed buyer, Message message, uint256 amountPaid, address token);
 
     event SetPrices(
-        uint256[] catalystIds,
+        uint16[] catalystIds,
         uint256[] catalystPrices,
-        uint256[] gemIds,
+        uint16[] gemIds,
         uint256[] gemPrices,
         uint256 priceChangeTimestamp
     );
 
     struct Message {
-        uint256[] catalystIds;
+        uint16[] catalystIds;
         uint256[] catalystQuantities;
-        uint256[] gemIds;
+        uint16[] gemIds;
         uint256[] gemQuantities;
         uint256 nonce;
     }
@@ -85,21 +85,21 @@ contract StarterPackV2 is PurchaseValidator, ERC2771Handler {
     /// @param gemIds Array of gem IDs for which new prices will take effect after a delay period
     /// @param gemPrices Array of new gems prices that will take effect after a delay period
     function setPrices(
-        uint256[] calldata catalystIds,
+        uint16[] calldata catalystIds,
         uint256[] calldata catalystPrices,
-        uint256[] calldata gemIds,
+        uint16[] calldata gemIds,
         uint256[] calldata gemPrices
     ) external onlyRole(DEFAULT_ADMIN_ROLE) {
         require(catalystIds.length == catalystPrices.length, "INVALID_CAT_INPUT");
         require(gemIds.length == gemPrices.length, "INVALID_GEM_INPUT");
         for (uint256 i = 0; i < catalystIds.length; i++) {
-            uint16 id = uint16(catalystIds[i]);
+            uint16 id = catalystIds[i];
             require(_isValidCatalyst(id), "INVALID_CAT_ID");
             _catalystPreviousPrices[id] = _catalystPrices[id];
             _catalystPrices[id] = catalystPrices[i];
         }
         for (uint256 i = 0; i < gemIds.length; i++) {
-            uint16 id = uint16(gemIds[i]);
+            uint16 id = gemIds[i];
             require(_isValidGem(id), "INVALID_GEM_ID");
             _gemPreviousPrices[id] = _gemPrices[id];
             _gemPrices[id] = gemPrices[i];
@@ -114,18 +114,18 @@ contract StarterPackV2 is PurchaseValidator, ERC2771Handler {
     /// @param gemIds The IDs of the gems to be transferred
     function withdrawAll(
         address to,
-        uint256[] calldata catalystIds,
-        uint256[] calldata gemIds
+        uint16[] calldata catalystIds,
+        uint16[] calldata gemIds
     ) external onlyRole(DEFAULT_ADMIN_ROLE) {
         for (uint256 i = 0; i < catalystIds.length; i++) {
-            uint16 id = uint16(catalystIds[i]);
+            uint16 id = catalystIds[i];
             require(_isValidCatalyst(id), "INVALID_CATALYST_ID");
             ICatalyst catalyst = _getCatalyst(id);
             uint256 balance = catalyst.balanceOf(address(this));
             _executeRegistryTransferCatalyst(catalyst, address(this), to, balance);
         }
         for (uint256 i = 0; i < gemIds.length; i++) {
-            uint16 id = uint16(gemIds[i]);
+            uint16 id = gemIds[i];
             require(_isValidGem(id), "INVALID_GEM_ID");
             IGem gem = _getGem(id);
             uint256 balance = gem.balanceOf(address(this));
@@ -178,7 +178,7 @@ contract StarterPackV2 is PurchaseValidator, ERC2771Handler {
     /// @return gemPricesBeforeSwitch Gem prices before price change
     /// @return gemPricesAfterSwitch Gem prices after price change
     /// @return switchTime The time the latest price change will take effect, being the time of the price change plus the price change delay
-    function getPrices(uint256[] calldata catalystIds, uint256[] calldata gemIds)
+    function getPrices(uint16[] calldata catalystIds, uint16[] calldata gemIds)
         external
         view
         returns (
@@ -198,12 +198,12 @@ contract StarterPackV2 is PurchaseValidator, ERC2771Handler {
         uint256[] memory gemPricesBeforeSwitch = new uint256[](gemIds.length);
         uint256[] memory gemPricesAfterSwitch = new uint256[](gemIds.length);
         for (uint256 i = 0; i < catalystIds.length; i++) {
-            uint16 id = uint16(catalystIds[i]);
+            uint16 id = catalystIds[i];
             catalystPricesBeforeSwitch[i] = _catalystPreviousPrices[id];
             catalystPricesAfterSwitch[i] = _catalystPrices[id];
         }
         for (uint256 i = 0; i < gemIds.length; i++) {
-            uint16 id = uint16(gemIds[i]);
+            uint16 id = gemIds[i];
             gemPricesBeforeSwitch[i] = _gemPreviousPrices[id];
             gemPricesAfterSwitch[i] = _gemPrices[id];
         }
@@ -235,33 +235,33 @@ contract StarterPackV2 is PurchaseValidator, ERC2771Handler {
     /// @param gemQuantities An array of gem amounts to be purchased
     /// @return the total price to pay in SAND for the cats and gems in the bundle
     function calculateTotalPriceInSAND(
-        uint256[] memory catalystIds,
+        uint16[] memory catalystIds,
         uint256[] memory catalystQuantities,
-        uint256[] memory gemIds,
+        uint16[] memory gemIds,
         uint256[] memory gemQuantities
     ) external returns (uint256) {
         return _calculateTotalPriceInSAND(catalystIds, catalystQuantities, gemIds, gemQuantities);
     }
 
     function _transferCatalysts(
-        uint256[] memory catalystIds,
+        uint16[] memory catalystIds,
         uint256[] memory catalystQuantities,
         address buyer
     ) internal {
         for (uint256 i = 0; i < catalystIds.length; i++) {
-            uint16 id = uint16(catalystIds[i]);
+            uint16 id = catalystIds[i];
             require(_isValidCatalyst(id), "INVALID_CATALYST_ID");
             _executeRegistryTransferCatalyst(_getCatalyst(id), address(this), buyer, catalystQuantities[i]);
         }
     }
 
     function _transferGems(
-        uint256[] memory gemIds,
+        uint16[] memory gemIds,
         uint256[] memory gemQuantities,
         address buyer
     ) internal {
         for (uint256 i = 0; i < gemIds.length; i++) {
-            uint16 id = uint16(gemIds[i]);
+            uint16 id = gemIds[i];
             require(_isValidGem(id), "INVALID_GEM_ID");
             _executeRegistryTransferGem(_getGem(id), address(this), buyer, gemQuantities[i]);
         }
@@ -303,9 +303,9 @@ contract StarterPackV2 is PurchaseValidator, ERC2771Handler {
 
     /// @dev Function to calculate the total price in SAND of the StarterPacks to be purchased
     function _calculateTotalPriceInSAND(
-        uint256[] memory catalystIds,
+        uint16[] memory catalystIds,
         uint256[] memory catalystQuantities,
-        uint256[] memory gemIds,
+        uint16[] memory gemIds,
         uint256[] memory gemQuantities
     ) internal returns (uint256) {
         require(catalystIds.length == catalystQuantities.length, "INVALID_CAT_INPUT");
@@ -313,14 +313,14 @@ contract StarterPackV2 is PurchaseValidator, ERC2771Handler {
         bool useCurrentPrices = _priceSelector();
         uint256 totalPrice;
         for (uint256 i = 0; i < catalystIds.length; i++) {
-            uint16 id = uint16(catalystIds[i]);
+            uint16 id = catalystIds[i];
             uint256 quantity = catalystQuantities[i];
             totalPrice =
                 totalPrice +
                 (useCurrentPrices ? _catalystPrices[id] * (quantity) : _catalystPreviousPrices[id] * (quantity));
         }
         for (uint256 i = 0; i < gemIds.length; i++) {
-            uint16 id = uint16(gemIds[i]);
+            uint16 id = gemIds[i];
             uint256 quantity = gemQuantities[i];
             totalPrice =
                 totalPrice +
